requirements.txt
yfinance
python-telegram-bot==20.3
openai
schedule